Rename controls to common.
