+++ front/src/common/flawless-victory.scss
