+++ front/src/common/flawless-victory.tsx
