+++ front/src/common/lifes.scss
