+++ front/src/common/lifes.tsx
