+++ front/src/common/question.scss
