+++ front/src/common/remaining_questions.css
